requirements.txt
fastapi~=0.112.2
pydantic~=2.8.2
requests~=2.32.3
uvicorn[standard]
azure-storage-blob